Implement base classes for players and matchings.
