+++ src/matching/__init__.py
@@ -7,9 +7,18 @@ if not sys.warnoptions:
 
     warnings.simplefilter("always")
 
-from .game import BaseGame
-from .matching import Matching
-from .player import Player
+from .base import BaseGame, BaseMatching, BasePlayer
+from .matchings import MultipleMatching, SingleMatching
+from .players import Player
 from .version import __version__
 
-__all__ = [BaseGame, Matching, Player, __version__]
+__all__ = [
+    "BaseGame",
+    "BaseMatching",
+    "BasePlayer",
+    "Matching",
+    "MultipleMatching",
+    "Player",
+    "SingleMatching",
+    "__version__",
+]
