+++ src/matching/players/player.py
@@ -1,7 +1,9 @@
 """ The base Player class for use in various games. """
 
+from matching import BasePlayer
 
-class Player:
+
+class Player(BasePlayer):
     """ A class to represent a player within the matching game.
 
     Parameters
@@ -23,36 +25,6 @@ class Player:
         The original set of player preferences.
     """
 
-    def __init__(self, name):
-
-        self.name = name
-        self.prefs = None
-        self.pref_names = None
-        self.matching = None
-        self._original_prefs = None
-
-    def __repr__(self):
-
-        return str(self.name)
-
-    def unmatched_message(self):
-
-        return f"{self} is unmatched."
-
-    def not_in_preferences_message(self, other):
-
-        return (
-            f"{self} is matched to {other} but they do not appear in their "
-            f"preference list: {self.prefs}."
-        )
-
-    def set_prefs(self, players):
-        """ Set the player's preferences to be a list of players. """
-
-        self.prefs = players
-        self.pref_names = [player.name for player in players]
-        self._original_prefs = players[:]
-
     def get_favourite(self):
         """ Get the player's favourite player. """
 
@@ -68,26 +40,11 @@ class Player:
 
         self.matching = None
 
-    def forget(self, other):
-        """ Forget another player by removing them from the player's preference
-        list. """
-
-        prefs = self.prefs[:]
-        prefs.remove(other)
-        self.prefs = prefs
-
     def get_successors(self):
         """ Get all the successors to the current match of the player. """
 
         idx = self.prefs.index(self.matching)
         return self.prefs[idx + 1 :]
-
-    def prefers(self, player, other):
-        """ Determines whether the player prefers a player over some other
-        player. """
-
-        prefs = self._original_prefs
-        return prefs.index(player) < prefs.index(other)
 
     def check_if_match_is_unacceptable(self, unmatched_okay=False):
         """ Check the acceptability of the current match, with the stipulation
